requirements.txt
asgiref==3.6.0
Django==4.1.7
django-simple-history==3.3.0
django-tables2==2.5.3
sqlparse==0.4.3
tzdata==2023.3